refactor chat sidebar to be RSC shell + client leafs. Move loading state to use the RSC shell.
