+++ src/app/(private)/dashboard/chat/loading.tsx
@@ -1,5 +1,19 @@
-import { ChatRouteLoading } from "@/components/chat/chat-route-loading";
+import {
+  ChatContentPane,
+  ChatSidebarPane,
+} from "@/components/chat/chat-layout-panels";
+import { ChatThreadLoading } from "@/components/chat/chat-loading";
+import { ChatSidebarLoading } from "@/components/chat/chat-sidebar";
 
 export default function ChatLoading() {
-  return <ChatRouteLoading />;
+  return (
+    <div className="flex h-full min-h-0 w-full">
+      <ChatSidebarPane>
+        <ChatSidebarLoading />
+      </ChatSidebarPane>
+      <ChatContentPane>
+        <ChatThreadLoading />
+      </ChatContentPane>
+    </div>
+  );
 }
